+++ src/main/java/tornadofx/Nodes.kt
@@ -663,6 +663,9 @@ class HBoxConstraint(
     }
 }
 
+var Node.hgrow: Priority get() = HBox.getHgrow(this); set(value) { HBox.setHgrow(this, value) }
+var Node.vgrow: Priority get() = VBox.getVgrow(this); set(value) { VBox.setVgrow(this, value) }
+
 fun <T : Node> T.anchorpaneConstraints(op: AnchorPaneConstraint.() -> Unit): T {
     val c = AnchorPaneConstraint()
     c.op()