+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.5.8-SNAPSHOT]
 
 - ItemViewModel().bindTo(listCellFragment)
+- resources.stream("some-resource") locates InputStream for resource
 
 ## [1.5.7] - 2016-10-21